refactor: move the Cairo 1 hint processor into its own crate
The cairo-vm library depended on cairo-lang-casm and cairo-lang-starknet-classes
behind the cairo-1-hints feature, which pulled the whole Cairo 1 compiler into
the dependency tree of anyone who only wanted to run Cairo 0 programs.

Executing Cairo 1 hints means speaking the compiler's Hint AST, so rather than
copying those definitions into cairo-vm, the hint processor moves to a new
cairo1-hint-processor crate that owns the compiler dependency. cairo-vm is now
free of cairo-lang-* in every dependency kind, including dev, and the
cairo-1-hints feature is gone rather than merely emptied.

The move needed no new public surface on cairo-vm: the tests that came along
switched from the private CairoRunner::program field to the existing
get_program()/data_len() accessors.

TryFrom<CasmContractClass> for Program could not come along as an impl, since
both types are foreign to the new crate; it is now the free function
program_from_casm_contract_class. Nothing outside cairo-vm's own tests called
it, so nothing is stranded.

cairo1-run needed only an import swap - it already produced
cairo_lang_casm::hints::Hint and the hint processor still consumes exactly that
type, so no conversion layer exists anywhere.
