+++ Cargo.toml
@@ -2,12 +2,13 @@
 members = [
     "cairo-vm-cli",
     "vm",
+    "cairo1-hint-processor",
     "hint_accountant",
     "cairo1-run",
     "cairo-vm-tracer",
     "examples/hyper_threading",
 ]
-default-members = ["cairo-vm-cli", "vm", "cairo1-run"]
+default-members = ["cairo-vm-cli", "vm", "cairo1-hint-processor", "cairo1-run"]
 
 # Explicitly set the resolver to the default for edition >= 2021
 # https://doc.rust-lang.org/edition-guide/rust-2021/default-cargo-resolver.html
@@ -24,6 +25,7 @@ keywords = ["starknet", "cairo", "vm"]
 [workspace.dependencies]
 cairo-vm = { path = "./vm", version = "3.2.0" }
 cairo-vm-tracer = { path = "./cairo-vm-tracer", version = "3.2.0" }
+cairo1-hint-processor = { path = "./cairo1-hint-processor", version = "3.2.0" }
 mimalloc = { version = "0.1.37", default-features = false }
 num-bigint = { version = "0.4", features = ["serde", "rand"] }
 rand = { version = "0.8.5", features = ["small_rng"] }
@@ -46,14 +48,13 @@ thiserror = { version = "2" }
 
 bitvec = { version = "1" }
 
-# Dependencies for cairo-1-hints feature
+# The Cairo 1 compiler. Only `cairo1-hint-processor` and the `cairo1-run` binary may depend on
+# these; `cairo-vm` itself must stay free of `cairo-lang-*`.
 cairo-lang-casm = { version = "2.12.0-dev.0" }
-
 cairo-lang-starknet-classes = { version = "2.12.0-dev.0" }
 cairo-lang-compiler = { version = "=2.12.0-dev.0" }
 cairo-lang-sierra-to-casm = { version = "2.12.0-dev.0" }
 cairo-lang-sierra = { version = "2.12.0-dev.0" }
-cairo-lang-runner = { version = "2.12.0-dev.0" }
 cairo-lang-utils = { version = "=2.12.0-dev.0" }
 
 ark-ff = { version = "0.4.2" }
