+++ cairo1-hint-processor/src/lib.rs
@@ -1,0 +1,23 @@
+//! # A `HintProcessor` for the hints introduced by the Cairo 1 compiler
+//!
+//! This lives outside the `cairo-vm` crate on purpose: executing Cairo 1 hints means speaking the
+//! compiler's `Hint` AST, so this is the only crate in the workspace that depends on
+//! `cairo-lang-*`. Running Cairo 0 programs pulls in none of it.
+//!
+//! ## Feature Flags
+//! - `extensive_hints`: forwards to [`cairo_vm`]'s feature of the same name, which lets a hint
+//!   extend the set of hints used for the rest of the run.
+
+#![deny(warnings)]
+#![forbid(unsafe_code)]
+
+pub mod circuit;
+pub mod dict_manager;
+pub mod hint_processor;
+pub mod hint_processor_utils;
+mod program;
+
+pub use program::program_from_casm_contract_class;
+
+#[cfg(test)]
+mod tests;
