+++ cairo1-run/Cargo.toml
@@ -9,7 +9,8 @@ keywords.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cairo-vm = { workspace = true, features = ["cairo-1-hints", "clap"] }
+cairo-vm = { workspace = true, features = ["clap"] }
+cairo1-hint-processor.workspace = true
 serde_json = { workspace = true }
 
 cairo-lang-sierra-type-size = { version = "2.12.0-dev.0", default-features = false }
