--- vm/src/hint_processor/cairo_1_hint_processor/mod.rs
@@ -1,4 +1,0 @@
-pub mod circuit;
-pub mod dict_manager;
-pub mod hint_processor;
-pub mod hint_processor_utils;
