+++ vm/src/hint_processor/mod.rs
@@ -1,5 +1,3 @@
 pub mod builtin_hint_processor;
-#[cfg(feature = "cairo-1-hints")]
-pub mod cairo_1_hint_processor;
 pub mod hint_processor_definition;
 pub mod hint_processor_utils;
