+++ vm/src/lib.rs
@@ -6,7 +6,6 @@
 //!    - the `print_*` family of hints;
 //!    - the `skip_next_instruction()` hints;
 //!    - implementations of [`arbitrary::Arbitrary`](https://docs.rs/arbitrary/latest/arbitrary/) for some structs.
-//! - `cairo-1-hints`: Enable hints that were introduced in Cairo 1. Not enabled by default.
 //! - `cairo-0-secp-hints`: Enable secp hints that were introduced in Cairo 0. Not enabled by default.
 //! - `cairo-0-data-availability-hints`: Enable data availability hints that were introduced in Cairo 0. Not enabled by default.
