+++ vm/src/types/program.rs
@@ -11,8 +11,6 @@ use crate::{
     vm::runners::cairo_pie::StrippedProgram,
 };
 
-#[cfg(feature = "cairo-1-hints")]
-use crate::serde::deserialize_program::{ApTracking, FlowTrackingData};
 use crate::utils::PRIME_STR;
 use crate::Felt252;
 use crate::{
@@ -25,8 +23,6 @@ use crate::{
         errors::program_errors::ProgramError, instruction::Register, relocatable::MaybeRelocatable,
     },
 };
-#[cfg(feature = "cairo-1-hints")]
-use cairo_lang_starknet_classes::casm_contract_class::CasmContractClass;
 use core::num::NonZeroUsize;
 
 use std::path::Path;
@@ -417,52 +413,6 @@ impl Default for Program {
             constants: Arc::new(HashMap::new()),
             builtins: Vec::new(),
         }
-    }
-}
-
-#[cfg(feature = "cairo-1-hints")]
-// Note: This Program will only work when using run_from_entrypoint, and the Cairo1Hintprocesso
-impl TryFrom<CasmContractClass> for Program {
-    type Error = ProgramError;
-    fn try_from(value: CasmContractClass) -> Result<Self, ProgramError> {
-        let data = value
-            .bytecode
-            .iter()
-            .map(|x| MaybeRelocatable::from(Felt252::from(&x.value)))
-            .collect();
-        //Hint data is going to be hosted processor-side, hints field will only store the pc where hints are located.
-        // Only one pc will be stored, so the hint processor will be responsible for executing all hints for a given pc
-        let hints = value
-            .hints
-            .iter()
-            .map(|(x, _)| {
-                (
-                    *x,
-                    vec![HintParams {
-                        code: x.to_string(),
-                        accessible_scopes: Vec::new(),
-                        flow_tracking_data: FlowTrackingData {
-                            ap_tracking: ApTracking::default(),
-                            reference_ids: HashMap::new(),
-                        },
-                    }],
-                )
-            })
-            .collect();
-        let error_message_attributes = Vec::new();
-        let reference_manager = ReferenceManager {
-            references: Vec::new(),
-        };
-        Self::new(
-            vec![],
-            data,
-            None,
-            hints,
-            reference_manager,
-            HashMap::new(),
-            error_message_attributes,
-            None,
-        )
     }
 }
